Osio 8 aloitettu

+++ Osio 4/Tehtava_3.py
@@ -2,15 +2,15 @@ x = input("Anna luku: ")
 if x == "":
     print("Toiminta lopetettu!")
 else:
-    min = int(x)
-    max = int(x)
+    min = float(x)
+    max = float(x)
 
 
 while x != "":
-    if int(x) < min:
-        min = int(x)
-    elif int(x) > max:
-        max = int(x)
+    if float(x) < min:
+        min = float(x)
+    elif float(x) > max:
+        max = float(x)
     x = input("Anna luku: ")
 
 print(f"Toiminta lopetettu! Pienin annettu luku oli {str(min):3} ja suurin annettu luku oli {str(max):3}.")
